create ViewModels
