+++ eMyAppTeDu.sln
@@ -14,7 +14,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "eShopSolution.Data", "eShop
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "eShopSolution.Application", "eShopSolution.Application\eShopSolution.Application.csproj", "{C2F8EA6F-EA84-4D08-A1C4-951512194BE6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "eShopSolution.Utilities", "eShopSolution.Utilities\eShopSolution.Utilities.csproj", "{D6E37141-34D9-4151-8DE4-ACFC4709F4D0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "eShopSolution.Utilities", "eShopSolution.Utilities\eShopSolution.Utilities.csproj", "{D6E37141-34D9-4151-8DE4-ACFC4709F4D0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "eShopSolution.ViewModels", "eShopSolution.ViewModels\eShopSolution.ViewModels.csproj", "{591F5230-31EB-47CC-8441-7B1BFD6316C1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,6 +40,10 @@ Global
 		{D6E37141-34D9-4151-8DE4-ACFC4709F4D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D6E37141-34D9-4151-8DE4-ACFC4709F4D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D6E37141-34D9-4151-8DE4-ACFC4709F4D0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{591F5230-31EB-47CC-8441-7B1BFD6316C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{591F5230-31EB-47CC-8441-7B1BFD6316C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{591F5230-31EB-47CC-8441-7B1BFD6316C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{591F5230-31EB-47CC-8441-7B1BFD6316C1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
