+++ eShopSolution.Application/Catalog/Products/IManageProductService.cs
@@ -1,9 +1,6 @@
-﻿using eShopSolution.Application.Catalog.DTOS;
-using eShopSolution.Application.Catalog.Products.Dtos;
-using eShopSolution.Application.Catalog.Products.Dtos.Manage;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using eShopSolution.ViewModels.Catalog.Products;
+using eShopSolution.ViewModels.Catalog.Products.Manage;
+using eShopSolution.ViewModels.Common;
 using System.Threading.Tasks;
 
 namespace eShopSolution.Application.Catalog.Products
