+++ eShopSolution.ViewModels/Catalog/Products/Manage/GetProductPagingRequest.cs
@@ -1,9 +1,9 @@
-﻿using eShopSolution.Application.Catalog.DTOS;
+﻿using eShopSolution.ViewModels.Common;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace eShopSolution.Application.Catalog.Products.Dtos.Manage
+namespace eShopSolution.ViewModels.Catalog.Products.Manage
 {
     public class GetProductPagingRequest : PagingRequestBase
     {
