+++ eShopSolution.ViewModels/Catalog/Products/Manage/ProductUpdateRequest.cs
@@ -2,16 +2,11 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace eShopSolution.Application.Catalog.Products.Dtos
+namespace eShopSolution.ViewModels.Catalog.Products.Manage
 {
-    public class ProductViewModel
+    public class ProductUpdateRequest
     {
-        public int Id { set; get; }
-        public decimal Price { set; get; }
-        public decimal OriginalPrice { set; get; }
-        public int Stock { set; get; }
-        public int ViewCount { set; get; }
-        public DateTime DateCreated { set; get; }
+        public int Id { get; set; }
         public string Name { set; get; }
         public string Description { set; get; }
         public string Details { set; get; }
