+++ eShopSolution.ViewModels/Catalog/Products/ProductViewModel.cs
@@ -2,11 +2,16 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace eShopSolution.Application.Catalog.Products.Dtos.Manage
+namespace eShopSolution.ViewModels.Catalog.Products
 {
-    public class ProductUpdateRequest
+    public class ProductViewModel
     {
-        public int Id { get; set; }
+        public int Id { set; get; }
+        public decimal Price { set; get; }
+        public decimal OriginalPrice { set; get; }
+        public int Stock { set; get; }
+        public int ViewCount { set; get; }
+        public DateTime DateCreated { set; get; }
         public string Name { set; get; }
         public string Description { set; get; }
         public string Details { set; get; }
