+++ eShopSolution.ViewModels/Catalog/Products/Public/GetProductPagingRequest.cs
@@ -1,0 +1,12 @@
+﻿using eShopSolution.ViewModels.Common;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace eShopSolution.ViewModels.Catalog.Products.Public
+{
+    public class GetProductPagingRequest : PagingRequestBase
+    {
+        public int CategoryId { get; set; }
+    }
+}
